OP-322: Move traits to Trait folder, implement getContent() method
