+++ src/Entity/ContentConfiguration.php
@@ -10,6 +10,8 @@ declare(strict_types=1);
 
 namespace BitBag\SyliusCmsPlugin\Entity;
 
+use BitBag\SyliusCmsPlugin\Form\Type\ContentElements\TextareaContentElementType;
+
 class ContentConfiguration implements ContentConfigurationInterface
 {
     protected ?int $id;
@@ -65,5 +67,14 @@ class ContentConfiguration implements ContentConfigurationInterface
     public function setPage(?PageInterface $page): void
     {
         $this->page = $page;
+    }
+
+    public function getContent(): ?string
+    {
+        if (TextareaContentElementType::TYPE === $this->type) {
+            return $this->configuration[TextareaContentElementType::TYPE] ?? null;
+        }
+
+        return null;
     }
 }
