+++ src/Entity/Media.php
@@ -10,6 +10,9 @@ declare(strict_types=1);
 
 namespace BitBag\SyliusCmsPlugin\Entity;
 
+use BitBag\SyliusCmsPlugin\Entity\Trait\ChannelsAwareTrait;
+use BitBag\SyliusCmsPlugin\Entity\Trait\CollectibleTrait;
+use BitBag\SyliusCmsPlugin\Entity\Trait\ProductsAwareTrait;
 use BitBag\SyliusCmsPlugin\MediaProvider\FilenameHelper;
 use Sylius\Component\Resource\Model\ToggleableTrait;
 use Sylius\Component\Resource\Model\TranslatableTrait;
