+++ src/Entity/Trait/ChannelsAwareTrait.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
  * We are hiring developers from all over the world. Join us and start your new, exciting adventure and become part of us: https://bitbag.io/career
 */
 
-namespace BitBag\SyliusCmsPlugin\Entity;
+namespace BitBag\SyliusCmsPlugin\Entity\Trait;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
