--- src/Form/Type/BlockImageType.php
@@ -1,21 +1,0 @@
-<?php
-
-/*
- * This file was created by developers working at BitBag
- * Do you need more information about us and what we do? Visit our https://bitbag.io website!
- * We are hiring developers from all over the world. Join us and start your new, exciting adventure and become part of us: https://bitbag.io/career
-*/
-
-declare(strict_types=1);
-
-namespace BitBag\SyliusCmsPlugin\Form\Type;
-
-use Sylius\Bundle\CoreBundle\Form\Type\ImageType as BaseImageType;
-
-final class BlockImageType extends BaseImageType
-{
-    public function getBlockPrefix(): string
-    {
-        return 'bitbag_sylius_cms_plugin_block_image';
-    }
-}
